[DoctrineBridge] Add LazyEventManager
